[ refactor ] Move `RedEx` things to `redex.rs`
